debugging subspace projection on gpu

+++ replicate-dcfnet/Projection/regularization.py
@@ -59,7 +59,7 @@ def subspace_projection_gpu(k,w,basis=None, basis_indices=None):
     for ii in range(F):
         indices = basis_indices[ii,:].astype(int)
         for jj in range(C):
-            basis_i = basis[:,indices]
+            basis_i = basis[:,torch.LongTensor(indices)]
             #B = np.dot(basis[:,indices],basis[:,indices].T)
             B = torch.mm(basis_i,torch.transpose(basis_i,0,1))
             #w_projected[ii,jj,:] = np.dot(B,w_reshaped[ii,jj,:])
